+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Add `Set.Empty()` static method.
 
 ## [2.2.0] - 2018-03-27
 - Add `Enumerables.ToDictionary` extension methods.